requirements.txt
# Core Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0

# Metrics and Evaluation
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0

# Progress Bars and Logging
tqdm>=4.65.0

# Optional: Word embeddings
gensim>=4.3.0

# Visualization (optional, for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (optional, for exploration)
jupyter>=1.0.0
ipykernel>=6.25.0

# XML parsing (for BC5CDR dataset)
lxml>=4.9.0
beautifulsoup4>=4.12.0

# Testing
pytest>=7.4.0

# Web Demo (optional)
gradio>=4.0.0
